+++ src/stores/index.js
@@ -1,0 +1,1 @@
+export { default as authStore } from './authStore';
